refactor tests + OptionChoice tests
